lnwallet: add ephemeral route blinding point to `PaymentDescriptor`
This places the route blinding point in our in-memory update log,
so it is available for use during onion & route blinding payload
decryption for nodes inside a blinded route.

The blinding point may also be used to determine how errors, both
local and downstream, will be processed.

A forwarding node has two channel links involved in forwarding an HTLC,
both of which have their own update logs, payment descriptors, &
channel state machines. When an HTLC is forwarded, it is received and
added to a remote update log by the incoming link and added to a local
update log prior to commitment by the outgoing link. If we would like
the channel state machine to preserve whether or not the HTLC had an
ephemeral blinding point, then the blinding point must be looked
after in both locations.

TODO(11/21/22):
- When processing a failure, the HTLC flows through the switch in
the opposite direction. We need to decide which link will handle
determining how errors are to be forwarded.

Consider HTLC update retransmission (in-memory and from disk) code path.
It seems to me we have two types of retransmission:
- In memory retransmission (ie: our node had an internal error processing the HTLC update, or our peer restarted)
- From disk retransmission (ie: our link/switch/node restarted - power outage, crashed due to bug, etc.)

We will need to make sure that HTLC updates (adds, settle, fails) for
blind hops are processed correctly after a restart.