hop/iterator: use all-zero 32 byte onion HMAC as indicator of final hop
The ChannelLink’s decision about whether it is processing a final/exit hop used to be plugged (almost) directly into the sphinx package’s observation of 32 all 0x00-bytes for the onion HMAC. Somewhere along the line the condition which defines whether a node is the final/exit hop appears to have changed from 32 0x00-bytes for the onion HMAC to an empty short_channel_id in the top level onion TLV payload. It looks like things may have changed w.r.t determining the final/exit hop when moving TLV onion payloads?
- https://github.com/lightningnetwork/lnd/blob/23cc8389f2e7db968e859f2ee2426b0906c2dd5d/htlcswitch/iterator_test.go#L78-L79

We will return to the old days where the sphinx implementation encountering an all-zero 32 byte onion HMAC signals we are the final hop. We can no longer rely on the presence of a next_hop in the top level onion TLV payload as it will not be there for blinded hops.