htlcswitch/mock: allow `mockHopIterator` to handle both normal and blind hops
We expand the capability of the mockHopIterator such that it is able to
handle both normal and blind hops. The destinction is made via a check
for sentinel value which delineates the byte boundary between serialized
hops.

The implemenation is functional yet somewhat propietary in the sense that
it is specific to handling blind hops and not as general as it could be.
This lack of generality leaves the door open to eventual inclusion of the
full blown TLV serialization scheme into the mock iterator.

Without this or something like TLV we do not have a way to know
whether we should or even how to deserialize a variable length field like
the route blinding payload.