lnwallet: persist/save the route blinding point in the channel state machine
- add to previous LN wallet commit
- TODO: consider the persistence case where node undergoes restart and needs to reprocess HTLCs coming from disk.

If one looks at channeldb.LogUpdate, it seems that the entire lnwire.Message is persisted to disk.
This would include both the onion blob (which contains the blinding point for introduction node)
and the TLV extension (which contains the blinding point for intermediate nodess). I think this
means that we get reprocessing of blind hops after a restart for free!