lnwallet: add route blinding point to LogUpdate we create for remote updates we haven't signed yet
Ensure that the LogUpdate we create to represent pending remote updates (ie: those
updates for which we have committed to locally but have not yet signed for) contains
the route blinding point if available.

We have already completed the critical first step of ensuring that when an ADD
update arrives, the payment descriptor which represents it in our in-memory update
log contains the route blinding point. Let's be sure we thread that blinding point
through to the LogUpdate so it is written to disk!

This ensures that the blinding point survives should we restart in the middle of
a commitment dance where the incoming HTLC has been irrevocably added to our local
commitment transaction but has not yet been irrevocably commited to in our peer's
commitment transaction! I think :)