lnwallet: demonstrate can recover blinding point from pending remote update
NOTE: This works until Revocation, at which point the blinding point can
be restored from the LogUpdate in the forwarding package as demonstrated
previously. Is it whack to restore the blinding point in different ways?
Should it just be restored from the channeldb.HTLC?

Even in the case we are able to restore from channeldb.HTLC, I think we
still might want to be able to recover blinding points from pending
remote updates as those (will/will not?) yet have an HTLC on...?