multi: add representation of blinded paths
This commit introduces the concept of a blinded payment, which
consists of a blinded path, aggregate parameters and minimal
constraints for the path and basic/sane validation.

Payment relay info and payment constraints are placed in the
record package, because when the time comes to add support for
forwarding of blinded payments, we'll extract these structs
as TLVs in our encrypted onion blob. For now, this is not
important (it's just an easy way to represent this data),
but Aggregate(Relay/Constraints) aliases are used in the
routing package to emphasize that these a blinded payment uses
aggregate versions of the per-hop struct.