switchrpc: add SendOnion rpc
Add RPC for dispatching payments via onions. The payment
route and onion are computed by the caller and the onion
is delivered to the server for forwarding.

NOTE: The server does NOT process or peel the onion so it
assumed that the onion will be constructed such that the first
hop is encrypted to one of the server's channel partners.