temp: disable forwarding elligibility checks in SendOnion
We'll instead rely on a simple mapping from public key to
short channel ID. To make the fowarding elligibility checks
here is redunant given the checks that the Switch will make
internally during forwarding.

NOTE: This mapping is only needed if we want to allow clients to
specify the first hop by public key. If we add the ability to
specify the channel via ID to the RPC request, then we don't need
to consult this mapping.