routing: add route transformation as functional option
Allow configurable post-processing which transforms paths or
routes built by the ChannelRouter. This allows a setup in which
the ChannelRouter runs remotely from the HTLCSwitch to build
routes as if they we're constructed from the perspective of the
lnd instance itself by trimming hops. This means that the onions
the ChannelRouter builds will be ready for direct delivery to the
peer of an lnd backend via the SendOnion style RPC, without
requiring it be peeled.