multi: make result tracking mode configurable
Allow node configuraiton to specify whether to treat
HTLC attempts in the normal way or as being tracked
by a remote entity. This has the practical effect of
disabling the automatic cleanup of network result state
within the Switch. Instead, state deletion must be completed
via RPC.

This is an all or nothing designation until we can determine
whether it can be set on a per-HTLC attempt basis.