switchrpc: add TrackOnion rpc
Add RPC to lookup the status of a previously forwarded
onion. Allow callers of the TrackOnion rpc to indicate
whether they would like to handle errors themselves or
delegate error decryption to the server.

We take care to return ErrPaymentIDNotFound across RPC
boundary to the RPC caller. This will allow the caller
of TrackOnion to explicitly confirm that there is no HTLC
in-flight for the supplied attempt ID, so it is free to
safely re-attempt the payment.