routerrpc: allow BuildRoute RPC to use arbitrary source
The routerrpc.Server and ChannelRouter assume that the source
public key for a route is always the node’s public key when calling
BuildRoute.

We update to allow overriding of this field to any arbitrary public
key. The list of hop public keys must be connected to the source
via channels in our graph view.