channeldb: add means to acknowledge attempt by id
This allows an attempt to be marked as successfully
sent by the caller and is useful for resolving ambiguity
on startup as to whether an HTLC was succesfully sent
or whether it still needs to be sent.

The ambiguity arises when our HTLC dispatching process
plays out via RPC over the netowrk rather than in-process.