routing: acknowledge successful htlc dispatch
The ChannelRouter needs to distinguish on startup
between an attempt that was initiated successfully
and an attempt for which it is not known whether it
initiated successfully. It must attempt to track the
result only for attempts which are known to have been
initiated successfully. Otherwise, since we base retry
decision of the result of attempt tracking, we risk
duplicate attempts being made if SendHTLC is implemented
via RPC between two processes communicating over async
network.

This persisting of the acknowledgement from the Switch
server (local or remote) can be done on a best effort basis.

NOTE: The acknowledgement is not needed for correct function
when the Router + Switch run in the same process.