htlcswitch: prevent duplicates in SendHTLC
The underlying SendHTLC method in the Switch is modified
to be idempotent for a given payment attempt ID. This is
accomplished by persisting the intent to dispatch an HTLC
via an InitAttempt method *before* sending it to the network.

Subsequent calls with the same attempt ID will be safely
rejected, providing the remote router (payment life-cycle
manager) with an explicit and unambiguous acknowledgement
of the HTLC's status. This extends the duplicate safety
already provided by the Switch’s CircuitMap.

The Switch will not forward an onion with the same attempt
ID twice without the result for a given ID having been
cleaned from the Switch's network result store.