htlcswitch: prevent duplicates in SendHTLC
We're seeing what benefit to upstream clients is
provided if the underlying SendHTLC implementatation
(and by extension the SendOnion RPC) will not forward
the same attempt ID twice without the result for a
given ID having been cleaned from the result store.

We accomplish the duplicate prevention using the
InitAttempt method of the Switch Store.