multi: add NextAttemptID to PaymentAttemptDispatcher interface
Allow the ChannelRouter to obtain the identifiers it uses for htlc
attempts from the entity which is already responsible for carrying
out htlc dispatching (eg: HTLCSwitch).

This subtle difference will appear uninteresting for traditional or
monolithic lnd deployments in which a single HTLC dispatching Router
and Switch run together in a single process. The utility will be more
readily apparent when Router and Switch are made to run in separate
processes and if we allow multiple entities to use the Switch to forward
HTLC attempts.

With central coordination of attempt ID generation, multiple payment
life-cycle managing entities can share the same flat ID space without
collisions.