routing: add configurable attempt resumer pattern
This allows for the local and remote router to provide
different implementations of this function.

This allows for the resolving of any ambiguity in the status
of an HTLC in the case where the ChannelRouter and Switch run
in separate processes and communicate over the network.

When SendHTLC returns it is expected that the Router know, without
any uncertainty whether the HTLC is in-flight (no error) or that
it is neither in flight nor will it ever be in flight. This is
tricky to guarantee when the communication happens over a network.

Instead we allow the server to resolve any uncertainty by calling
SendHTLC while resuming a payment prior to attempting to track
the result. If the HTLC was successfully received by the remote
Switch, then the Router will receive a duplicate HTLC error and
can proceed to tracking the attempt result like normal.