htlcswitch: enable namspace encoding in msb of attempt ID
This overloads the most significant byte of the uint64
htlc attempt id to encode a namespace. This provides a
level of attempt ID space separation.

NOTE: The raw attempt ID will be globally unique across
all namespaces. This is not a requirement and one could
envision a fully namespace isolated ID space in which ids
are permitted to be re-used - just must be unique on a per
namespace basis.