htlcswitch: most significant byte namespacing in NextAttemptID
NOTE: NextAttemptID is the only place the encoding occurs. Once
the uint64 attemptID is returned by NextAttemptID, it must be
treated as opaque by all downstream consumers.

Eventually CleanStore can be updated to perform per namespace
deletion such that it is safe to be used by multiple routers
or entities managing payment-lifecycle.