multi: make result tracking mode configurable
Allow node configuraiton to specify whether to treat
HTLC attempts in the normal way or as being tracked
by a remote entity. This has the practical effect of
disabling the automatic cleanup of network result state
within the Switch. Instead, state deletion must be completed
via RPC.

This is an all or nothing designation and is intended to allow
an external routing (payment life-cycle managing) entity to
dispatch HTLC attempts via lnd's Switch RPC server.

NOTE: It is only safe to allow a *single* router (local or
remote) to dispatch attempts via the Switch at a time.