htlcswitch: add ErrAttemptResultNotAvailable
No longer return ErrPaymentIDNotFound in all
cases where the final network result is not yet
back from the network.

This makes the API more "honest" in the sense that it
will accurately differentiate between an attempt which
has been initialied but which does not yet have a final
(settle/fail) result and an attempt which has never been
heard of.

It requires that consumers of the network result store be
updated handle the exposure of this new state.