switchrpc: refactor SendOnion
Move the durable persistence of intent (InitAttempt) to
the true boundary of the HTLC dispatch flow. This serves,
not only as the basis for duplicate safety, but also permits
other errors returned by the Switch RPC server to be interpreted
by clients without ambiguity as they can be sure that an
error received other than ErrDuplicateAdd is from the *first*
time the HTLC dispatch request was received by the server.

Without this, the client must be extremlely conservative in
handling errors in which it may issue a payment attempt retry.
It cannot receive an ErrSwitchExiting error or transient link
error and conclude with any certainty as to the status of whether
or not an htlc with that ID has been dispatched. The error
communicates "the switch is not able to forward right now". It
says nothing about whether some prior SendOnion request was
received and successfully processed.