htlcswitch: prevent FailAttempt for non-init attempts
Add some protection which prevents FailAttempt from
overwriting actual (settle/fail) results coming back
from the network. We only permit FailAttempt to be
used to transition from INIT --> FAILED attempt state.
This avoids the scenario where an HTLC is initialized
but never actually dispatched to the network.

If such an attempt is left in the initialized state,
then subscribing clients could be stuck forever waiting
on a result for the HTLC attempt which will never arrive.