htlcswitch: add ambiguous init err to attempt store
This allows the Switch to communicate to a caller whether
the status of an HTLC dispatch for a given attempt ID is
ambiguous. This can occur when an attempt to initialize an
attempt (via InitAttempt) itself fails. This is the key
we use to provide idempotence or protection against duplicate
processing so that we can offer "at most once" semantics
when processing htlc dispatch requests.

If this InitAttempt call itself fails, then we cannot
definitively know if the attempt is not in-flight, or if it
was successfully processed by a prior request.

Callers may need to be updated to explicitly handle this
new ambiguous scenario.