routing: handle ambiguous attempt initialization
Consider ErrAmbiguousAttemptInit from SendHTLC as a failure
of the dispatcher (switch) to definitively communicate the
state of the htlc dispatch request. The current expected use
of this signal is for DB error on attempt initialization
itself.

When the router runs within the same process as the dispatcher,
this could be considered a definitive acknowledgement that
the htlc is *not* in flight. However, when the router runs
remotely from the dispatcher and communicates via rpc using
retries, we cannot assume the same certainty about htlc status.
Our only definitive acknowledgement from the dispatcher is a `nil`
error indicating the attempt is in-flight or an `ErrDuplicateAdd`
indicating the htlc by this id was already successfully received
and initialized. In either case, it is safe for the router to
proceed to tracking the result of the dispatch attempt.