htlcswitch: add file marker to prevent router collision
To prevent silent loss of htlc attempt information
and resulting stuck payments, this commit introduces a
file based marker which indicates whether lnd has been
used with a remote router (switchrpc) configuration.

When lnd is compiled with the 'switchrpc' tag, a
'remote_router.marker' file is created. The switch will
refuse to start if lnd is built without the switchrpc
server while this file is present on the file system,
forcing the user to take manual action.

To switch back to local management, the user must manually
delete the marker file after ensuring no in-flight payments
are held by the remote router.

The safety check during Switch creation prevents start
up of the Switch if it contains attempt data from use
with a remote router. This avoids lnd's onboard channel
router from clearing state needed to track payment
life-cycle for payments that it did not inititate.

NOTE: This may be unnecessary once the Switch stores
are namespaced and can be shared by multiple routers
at once.