htlcswitch: add test to document on-chain resolution bug
This commit adds a new test, TestOnChainResolutionFailure,
which demonstrates a error misclassification in the htlcswitch.

The test shows that when an on-chain resolution with a nil/empty
error reason is read from the result store, it is misclassified by
the parseFailedPayment function. On master, this leads to an
incorrect error being returned to the ChannelRouter by the
GetAttemptResult function.

This test documents the current (buggy) on-chain resolution behavior
and is corrected in the following commit.