switchrpc: overhaul TrackOnion rpc proto
The new structure uses a top-level `oneof` to provide a
compile-time distinction between a successful payment
(preimage) and a failed one. Additional information on a
failed attempt can be found in FailureDetails.

We now also use a structured ForwardingFailure type for
communicating the failure index and wire message from
failures which occur during htlc forwarding downstream
in the route.